requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.5.18.1
PySocks==1.7.1
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
